fix error when deleting recipe with no image and change register response from plain text to json

+++ src/main/java/com/springtest/cookapi/infrastructure/services/recipe/RecipeServiceImpl.java
@@ -93,7 +93,9 @@ public class RecipeServiceImpl implements IRecipeService{
         }
 
         recipeRepository.deleteById(recipeId);
-        cloudinaryService.deleteImageFromCloudinary(publicId);
+        if (publicId != null) {
+            cloudinaryService.deleteImageFromCloudinary(publicId);
+        }
     }
